Rama estiven

+++ CrudDF3/Controllers/RolesController.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using CrudDF3.Models;
 
@@ -27,13 +26,14 @@ namespace CrudDF3.Controllers
         // GET: Roles/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
+            if (!id.HasValue)
             {
                 return NotFound();
             }
 
             var role = await _context.Roles
-                .FirstOrDefaultAsync(m => m.IdRol == id);
+                .FirstOrDefaultAsync(m => m.IdRol == id.Value);
+
             if (role == null)
             {
                 return NotFound();
@@ -49,11 +49,9 @@ namespace CrudDF3.Controllers
         }
 
         // POST: Roles/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("IdRol,NombreRol,DescripcionRol,EstadoRol")] Role role)
+        public async Task<IActionResult> Create([Bind("NombreRol,DescripcionRol,EstadoRol")] Role role)
         {
             if (ModelState.IsValid)
             {
@@ -67,12 +65,12 @@ namespace CrudDF3.Controllers
         // GET: Roles/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
+            if (!id.HasValue)
             {
                 return NotFound();
             }
 
-            var role = await _context.Roles.FindAsync(id);
+            var role = await _context.Roles.FindAsync(id.Value);
             if (role == null)
             {
                 return NotFound();
@@ -81,13 +79,11 @@ namespace CrudDF3.Controllers
         }
 
         // POST: Roles/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("IdRol,NombreRol,DescripcionRol,EstadoRol")] Role role)
+        public async Task<IActionResult> Edit(int? id, [Bind("IdRol,NombreRol,DescripcionRol,EstadoRol")] Role role)
         {
-            if (id != role.IdRol)
+            if (id.GetValueOrDefault() != role.IdRol.GetValueOrDefault())
             {
                 return NotFound();
             }
@@ -118,13 +114,14 @@ namespace CrudDF3.Controllers
         // GET: Roles/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
+            if (!id.HasValue)
             {
                 return NotFound();
             }
 
             var role = await _context.Roles
-                .FirstOrDefaultAsync(m => m.IdRol == id);
+                .FirstOrDefaultAsync(m => m.IdRol == id.Value);
+
             if (role == null)
             {
                 return NotFound();
@@ -136,21 +133,26 @@ namespace CrudDF3.Controllers
         // POST: Roles/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
+        public async Task<IActionResult> DeleteConfirmed(int? id)
         {
-            var role = await _context.Roles.FindAsync(id);
+            if (!id.HasValue)
+            {
+                return NotFound();
+            }
+
+            var role = await _context.Roles.FindAsync(id.Value);
             if (role != null)
             {
                 _context.Roles.Remove(role);
+                await _context.SaveChangesAsync();
             }
 
-            await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
-        private bool RoleExists(int id)
+        private bool RoleExists(int? id)
         {
-            return _context.Roles.Any(e => e.IdRol == id);
+            return id.HasValue && _context.Roles.Any(e => e.IdRol == id.Value);
         }
     }
 }
